requirements.txt
Flask==3.1.2
flask-cors==6.0.1
googlesearch-python==1.3.0
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==21.2.0